snowflake ids

+++ app/database/database_schema.py
@@ -5,7 +5,7 @@ metadata = sqlalchemy.MetaData()
 
 users = sqlalchemy.Table("users",
     metadata,
-    sqlalchemy.Column("id", sqlalchemy.Integer, primary_key=True),
+    sqlalchemy.Column("id", sqlalchemy.BigInteger, primary_key=True),
     sqlalchemy.Column("token", sqlalchemy.VARCHAR(32), nullable=False),
     sqlalchemy.Column("email", sqlalchemy.VARCHAR(255), nullable=False),
     sqlalchemy.Column("password", sqlalchemy.Binary),
@@ -15,7 +15,7 @@ users = sqlalchemy.Table("users",
 
 assets = sqlalchemy.Table("assets",
     metadata,
-    sqlalchemy.Column("id", sqlalchemy.Integer, primary_key=True),
+    sqlalchemy.Column("id", sqlalchemy.BigInteger, primary_key=True),
     sqlalchemy.Column("token", sqlalchemy.VARCHAR(32), nullable=False),
     sqlalchemy.Column("owner", sqlalchemy.VARCHAR(32), sqlalchemy.ForeignKey("users.id"), nullable=False),
     sqlalchemy.Column("data", sqlalchemy.dialects.postgresql.JSONB, nullable=False),
